+++ presto-rcfile/src/test/java/com/facebook/presto/rcfile/RcFileTester.java
@@ -1039,7 +1039,7 @@ public class RcFileTester
             return ((SqlVarbinary) value).getBytes();
         }
         else if (type.equals(DATE)) {
-            int days = ((SqlDate) value).getDays();
+            long days = ((SqlDate) value).getDays();
             LocalDate localDate = LocalDate.ofEpochDay(days);
             ZonedDateTime zonedDateTime = localDate.atStartOfDay(ZoneId.systemDefault());
